+++ src/main/java/net/jrodolfo/java_evolution/java03/shutdown_hooks/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Explanatory module for JVM shutdown hooks.
+ * Executable examples for JVM shutdown hooks.
  */
 package net.jrodolfo.java_evolution.java03.shutdown_hooks;